Add files via upload
In This MATLAB code, I tried to find a way to model the permutation of supports' stiffness coefficient in order to be able to find the Minimum "Rm" between different permuted states of the problem; Rm is needed in order to find Minimum Brace Area of each beam that would prevent buckling in columns and is acquired through solving the matrix written in Problem.pdf file. 
The challenge in this code was modelling the "Tridiagonal Matrix" with the permutations embedded in it.

Disclaimer: This is a Freelancer Project.